Contents/requirements.txt
google-adk>=0.1.0
pypdf>=3.13.0
termcolor==3.1.0  # required by ADK runners
rich>=13.0.0      # optional, used for nicer CLI output